phase10: main_loop narrowing-cast fix + panic-stub gen hardening
Three changes in one commit:

1. main_loop.c line 198 dropped a bogus `(uint8_t)` cast on
   Sound_LoadSND's return value.  The Ghidra ref stores the full 32
   bits to uRam000005f8 (a uint32_t); the cleanup pass-2 narrowing was
   a decomp error.  uRam000005f8 is later passed to FUN_80044394
   (Audio_FreeSND) which dereferences it -- truncating to u8 would
   have crashed the freer the moment it got past panic-stub coverage.

2. tools/gen_panic_stubs.py: scans BOTH platform/ AND src/ for real
   function definitions now (was platform/-only).  Without this, when
   a cleaned engine function got a real body in src/, the regen still
   emitted a panic stub for it and the link failed LNK2005.

   Also strips `#if 0 ... #endif` blocks before scanning -- src/
   cleaned files keep the Ghidra-ref audit copy under that guard, and
   the prior regex would mistake those for real impls (false
   positive) and incorrectly OMIT a panic stub the link still needed.

   Tightened the prototype regex to require `(args) {` (body opens),
   excluding `extern <type> name(...);` declarations.

3. snd_loader.c: audited Sound_LoadSND wire-up.  Enabling it
   transitively pulls Stream_OpenByName -> FUN_80015368 (the engine's
   failure-path infinite-loop trap) and Audio_ParseSND ->
   SpuIsTransferCompleted, both of which are panic-stubbed and spin
   the loop.  Reverted the engine-name to its legacy 'Audio_LoadSND'
   so the link does not pull it in until those dependencies are real.
   Documented the seam contract for future audio enablement.

+++ src/gameplay/main_loop.c
@@ -195,7 +195,7 @@ void V8_MainLoop(void)
         }
         Audio_BankSelect(mask);
 
-        uRam000005f8 = (uint8_t)Sound_LoadSND("Sounds\\Main.SND");
+        uRam000005f8 = Sound_LoadSND("Sounds\\Main.SND");
         uint32_t fnt = Font_LoadFNT("Misc\\Game.FNT");
         Font_PrepareGlyphs(0);
         Font_SelectBank(1);
